pyproject.toml
[project]
name = "doccrawl"
version = "0.1.0"
description = "Document crawling and archiving system"
readme = "README.md"
requires-python = ">=3.11"
license = {file = "LICENSE"}
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]

dependencies = [
    "logfire>=1.3.0",
    "pandas>=2.2.3",
    "playwright>=1.48.0",
    "psycopg2-binary>=2.9.10", 
    "pydantic>=2.6.1",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "PyYAML>=6.0.1",
    "scrapegraphai>=1.27.0",
]

[project.scripts]
doccrawl = "doccrawl.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/doccrawl"]